Documentation and bug fixes
- grouped ctype sources
- smooth mouse movement
- fixed wm memory leak
- fixed readdir not null-ending entry names
- fixed paint drawing outside of buffer
- added an icon for paint

I wish I could put the wallpaper in the disk image like paint's icon,
but it's complicated: if the disk image is too big, grub aligns its
module to 4MiB, so modules go over the 4MiB mapping of the kernel. It's
difficult to map more than those 4MiB because the physical memory
manager needs to setup its bitmap, and I can only put it after GRUB
modules, which ends up unmapped. And we can't map more pages with the
PMM allocating page tables.

+++ kernel/include/kernel/mouse.h
@@ -27,7 +27,7 @@
 #define MOUSE_UNUSED_B (1 << 6)
 
 typedef struct {
-	int32_t x, y;
+	float x, y;
 	bool left_pressed;
 	bool right_pressed;
 	bool middle_pressed;
